Security: enforce Discord runtime operator authorization
